pyproject.toml
[project]
name = "trendradar-ai"
version = "1.0.0"
description = "AI-powered trending news aggregation and analysis system using CrewAI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10,<3.14"
authors = [
    {name = "TrendRadar AI Team"}
]
keywords = ["trending", "news", "aggregation", "crewai", "ai-analysis", "notifications"]

dependencies = [
    "crewai>=0.80.0",
    "crewai-tools>=0.14.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "tenacity>=8.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]
mcp = [
    "fastmcp>=0.1.0",
]

[project.scripts]
trendradar = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
